% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRCC.R
\docType{data}
\name{ccRCC}
\alias{ccRCC}
\title{A Clear cell renal cell carcinoma (ccRCC) data}
\format{An object of class \code{data.frame} with 18458 rows and 32 columns.}
\usage{
data(ccRCC)
}
\description{
The ccRCC RNA sequencing data contains genes annotated by Ensembl.
The Ensembl ID was converted to the HGNC symbol by Biomart and kept the protein coding
genes only. The processed dataset contains 18,458 protein coding genes and 32
patients with paired FFPE and RNAlater samples.
}
\examples{
data(ccRCC)
}
\references{
Eikrem et al. (2016) PLoS ONE 11(2)
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4764764/}{PubMed})
}
\keyword{datasets}
