% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMIXnorm.R
\name{dpoi}
\alias{dpoi}
\title{modified log likelihood for Poisson distribution}
\usage{
dpoi(x, delta)
}
\arguments{
\item{x}{is a vector}

\item{delta}{is a constant Poisson mean}
}
\value{
loglikelihood of non-zero elements of x
}
\description{
only calculate the non-zero elements of x likelihood
}
