% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXnorm.R
\name{func_MIXnorm}
\alias{func_MIXnorm}
\title{Produce MIXnorm normalized expression matrix}
\usage{
func_MIXnorm(dat, max_iter = 20, tol = 0.01, appr = T)
}
\arguments{
\item{dat}{input raw read count matrix. dim of dat = J genes * I samples.}

\item{max_iter}{maximum number of iterations for the nested EM algorithm default is 20, recommend range (10, 50).}

\item{tol}{convergency criteria, default is 1e-2, recommend range (1e-5,1).}

\item{appr}{binary True of False, indicates if the approximate version of normalization should be used.}
}
\value{
A list contains the normalized expression matrix, proportion of expressed genes and probabilities of being expressed for all genes.
}
\description{
This function calls the MIXnorm function to obtain MLE of the mixture model,
then produces the normalized expression matrix.
}
