% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMIXnorm.R
\name{func_SMIXnorm}
\alias{func_SMIXnorm}
\title{Produce SMIXnorm normalized expression matrix}
\usage{
func_SMIXnorm(
  dat,
  max_iter = 15,
  tol = 0.01,
  log_file = "SMIXnorm.log",
  appr = T
)
}
\arguments{
\item{dat}{input raw read count matrix. dim of dat = J genes * I samples.}

\item{max_iter}{maximum number of iterations for the nested EM algorithm default is 15, recommend range (10, 50).}

\item{tol}{convergency criteria, default is 1e-2, recommend range (1e-5,1).}

\item{log_file}{file name that records the log. Default log file name is SMIXnorm.log.}

\item{appr}{binary True of False, indicates if the approximate version of normalization should be used.}
}
\value{
A list contains the normalized expression matrix, proportion of expressed genes and probabilities of being expressed for all genes.
}
\description{
This function calls the SMIXnorm function to obtain MLE of the mixture model,
then produces the normalized expression matrix.
}
