% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXnorm.R
\name{MIXnorm.mle}
\alias{MIXnorm.mle}
\title{Main function for the nested EM algorithm for MIXnorm.}
\usage{
MIXnorm.mle(dat, max_iter = 20, tol = 0.01)
}
\arguments{
\item{dat}{input raw read count matrix. dim(dat)=J genes * I samples.}

\item{max_iter}{maximum number of iterations for the nested EM algorithm default is 20, recommend range (10, 50).}

\item{tol}{convergency criteria, default is 1e-2, recommend range (1e-5,1).}
}
\value{
A list contains the MLE of all parameters.
}
\description{
Performs the nested EM algorithm to find the MLE of the mixture model.
}
