% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMIXnorm.R
\name{SMIXnorm.mle}
\alias{SMIXnorm.mle}
\title{Main function for the nested EM algorithm for SMIXnorm.}
\usage{
SMIXnorm.mle(dat, max_iter = 20, tol = 0.01)
}
\arguments{
\item{dat}{input raw read count matrix. dim(dat)=J genes * I samples.}

\item{max_iter}{maximum number of iterations for the nested EM algorithm default is 20, recommend range (10, 50).}

\item{tol}{convergency criteria, default is 1e-2, recommend range (1e-5,1).}
}
\value{
A list conatins the MLE of all parameters.
}
\description{
Performs the nested EM algorithm to find the MLE of the simplified mixture model.
}
