% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXnorm.R
\name{MIXnorm}
\alias{MIXnorm}
\title{Main function for the nested EM algorithm for MIXnorm.}
\usage{
MIXnorm(dat, max_iter = 15, tol = 0.01, log_file)
}
\arguments{
\item{dat}{input raw read count matrix. dim(dat)=J genes * I samples.}

\item{max_iter}{maximum number of iterations for the nested EM algorithm default is 15, recommend range (10, 50).}

\item{tol}{convergency criteria, default is 1e-2, recommend range (1e-5,1).}

\item{log_file}{file name that records the log.}
}
\value{
the MLE of all parameters
}
\description{
This function takes the raw data as input and find the MLE of the mixture model.
}
